refactor: cache lib
