+++ pkg/listener/new_result_test.go
@@ -15,7 +15,7 @@ func Test_ResultListener(t *testing.T) {
 	t.Run("Publish Result", func(t *testing.T) {
 		var called v1alpha2.PolicyReportResult
 
-		slistener := listener.NewResultListener(true, cache.New(0, 5*time.Minute), time.Now())
+		slistener := listener.NewResultListener(true, cache.NewInMermoryCache(0, 5*time.Minute), time.Now())
 		slistener.RegisterListener(func(_ v1alpha2.ReportInterface, r v1alpha2.PolicyReportResult, b bool) {
 			called = r
 		})
@@ -30,7 +30,7 @@ func Test_ResultListener(t *testing.T) {
 	t.Run("Ignore Delete Event", func(t *testing.T) {
 		var called bool
 
-		slistener := listener.NewResultListener(true, cache.New(0, 5*time.Minute), time.Now())
+		slistener := listener.NewResultListener(true, cache.NewInMermoryCache(0, 5*time.Minute), time.Now())
 		slistener.RegisterListener(func(_ v1alpha2.ReportInterface, r v1alpha2.PolicyReportResult, b bool) {
 			called = true
 		})
@@ -45,7 +45,7 @@ func Test_ResultListener(t *testing.T) {
 	t.Run("Ignore Added Results created before startup", func(t *testing.T) {
 		var called bool
 
-		slistener := listener.NewResultListener(true, cache.New(0, 5*time.Minute), time.Now())
+		slistener := listener.NewResultListener(true, cache.NewInMermoryCache(0, 5*time.Minute), time.Now())
 		slistener.RegisterListener(func(_ v1alpha2.ReportInterface, r v1alpha2.PolicyReportResult, b bool) {
 			called = true
 		})
@@ -60,7 +60,7 @@ func Test_ResultListener(t *testing.T) {
 	t.Run("Ignore CacheResults", func(t *testing.T) {
 		var called bool
 
-		rcache := cache.New(0, 5*time.Minute)
+		rcache := cache.NewInMermoryCache(0, 5*time.Minute)
 		rcache.Add(fixtures.FailPodResult.ID)
 
 		slistener := listener.NewResultListener(true, rcache, time.Now())
@@ -78,7 +78,7 @@ func Test_ResultListener(t *testing.T) {
 	t.Run("Early Return if Results are empty", func(t *testing.T) {
 		var called bool
 
-		rcache := cache.New(0, 5*time.Minute)
+		rcache := cache.NewInMermoryCache(0, 5*time.Minute)
 		rcache.Add(fixtures.FailPodResult.ID)
 
 		slistener := listener.NewResultListener(true, rcache, time.Now())
